Move `OreDeposit` to `libOPHD`
