+++ appOPHD/States/MapViewStateDraw.cpp
@@ -9,7 +9,8 @@
 #include "../Constants/UiConstants.h"
 #include "../StructureManager.h"
 #include "../Map/TileMap.h"
-#include "../MapObjects/OreDeposit.h"
+
+#include <libOPHD/MapObjects/OreDeposit.h>
 
 #include <NAS2D/Utility.h>
 #include <NAS2D/Renderer/Renderer.h>
