+++ appOPHD/UI/MineOperationsWindow.cpp
@@ -10,9 +10,10 @@
 #include "../Constants/UiConstants.h"
 #include "../StructureManager.h"
 #include "../TruckAvailability.h"
-#include "../MapObjects/OreDeposit.h"
 #include "../MapObjects/Structures/MineFacility.h"
 #include "../MapObjects/Structures/Warehouse.h"
+
+#include <libOPHD/MapObjects/OreDeposit.h>
 
 #include <NAS2D/Utility.h>
 #include <NAS2D/Renderer/Renderer.h>
