+++ appOPHD/appOPHD.vcxproj
@@ -190,7 +190,6 @@
     <ClCompile Include="Map\Tile.cpp" />
     <ClCompile Include="Map\TileMap.cpp" />
     <ClCompile Include="MapObjects\MapObject.cpp" />
-    <ClCompile Include="MapObjects\OreDeposit.cpp" />
     <ClCompile Include="MapObjects\Robot.cpp" />
     <ClCompile Include="MapObjects\Robots\Robodigger.cpp" />
     <ClCompile Include="MapObjects\Robots\Robodozer.cpp" />
@@ -296,7 +295,6 @@
     <ClInclude Include="Map\Tile.h" />
     <ClInclude Include="Map\TileMap.h" />
     <ClInclude Include="MapObjects\MapObject.h" />
-    <ClInclude Include="MapObjects\OreDeposit.h" />
     <ClInclude Include="MapObjects\Robots\Robodigger.h" />
     <ClInclude Include="MapObjects\Robots\Robodozer.h" />
     <ClInclude Include="MapObjects\Robots\Robominer.h" />
