+++ appOPHD/appOPHD.vcxproj.filters
@@ -93,9 +93,6 @@
     <ClCompile Include="MapObjects\MapObject.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="MapObjects\OreDeposit.cpp">
-      <Filter>Source Files\MapObjects</Filter>
-    </ClCompile>
     <ClCompile Include="MapObjects\Robot.cpp">
       <Filter>Source Files\MapObjects</Filter>
     </ClCompile>
@@ -405,9 +402,6 @@
       <Filter>Header Files\Map</Filter>
     </ClInclude>
     <ClInclude Include="MapObjects\MapObject.h">
-      <Filter>Header Files\MapObjects</Filter>
-    </ClInclude>
-    <ClInclude Include="MapObjects\OreDeposit.h">
       <Filter>Header Files\MapObjects</Filter>
     </ClInclude>
     <ClInclude Include="MapObjects\Robots\Robodigger.h">
