+++ libOPHD/MapObjects/OreDeposit.cpp
