+++ libOPHD/libOPHD.vcxproj
@@ -166,6 +166,7 @@
     <ClCompile Include="StorableResources.cpp" />
     <ClCompile Include="XmlSerializer.cpp" />
     <ClCompile Include="Map\MapCoordinate.cpp" />
+    <ClCompile Include="MapObjects\OreDeposit.cpp" />
     <ClCompile Include="Population\Morale.cpp" />
     <ClCompile Include="Population\Population.cpp" />
     <ClCompile Include="Population\PopulationPool.cpp" />
@@ -190,6 +191,7 @@
     <ClInclude Include="XmlSerializer.h" />
     <ClInclude Include="Map\MapCoordinate.h" />
     <ClInclude Include="Map\MapOffset.h" />
+    <ClInclude Include="MapObjects\OreDeposit.h" />
     <ClInclude Include="MapObjects\StructureType.h" />
     <ClInclude Include="Population\Morale.h" />
     <ClInclude Include="Population\Population.h" />
